modularize attachments download
